requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
Pillow==10.1.0
pydantic==2.5.0
pydantic-settings==2.0.3
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
asyncio==3.4.3